Move the refactoring logic from the code action into separate file
